+++ app/static/game.js
@@ -27,8 +27,8 @@ const SOUNDS = {
     pickup: new Audio("/static/sounds/pickup.wav"),
     floor_cleared: new Audio("/static/sounds/clear.wav"),
     player_hurt: new Audio("/static/sounds/hurt.wav"),
-    game_win: new Audio("/static/sounds/slash.wav"),
-    game_lose: new Audio("/static/sounds/slash.wav"),
+    game_win: new Audio("/static/sounds/win.wav"),
+    game_lose: new Audio("/static/sounds/lose.wav"),
     play_card: new Audio("/static/sounds/card.wav"),
     explosion: new Audio("/static/sounds/explosion.wav"),
     attack: new Audio("/static/sounds/slash.wav"),
