--- game_enums.py
@@ -1,7 +1,0 @@
-import enum
-
-class Dificulty(enum.Enum):
-    easy = "easy"
-    normal = "normal"
-    hard = "hard"
-    master = "master"
